cleanup pt. 1
