+++ simulations/distributed/kafkaConfig/__init__.py
@@ -1,0 +1,1 @@
+config = {'send_topic': 'test', 'producer_config': {'bootstrap_servers': 'localhost:9092', 'acks': 'all'}}
